test(core): split build sync tests
